fix(item): add a default image url
in order to not force users to have to upload when creating a new item, adding a default image to the Item model is the best solution.

Providing a value for the image preview in the component when an image is not on the item itself allows us to always have an image to display, while at the same time not taking up space in the database. As well as if we change our mind on the default image's url in the future, it's a 1 line change instead of a database update.

#INC-5312: Missing image
Fixes #2