fix(item): add a default image url
in order to not force users to have to upload when creating a new item, adding a default image to the Item model is the best solution.

Providing a value for the image when serializing the Item model to json from the backend when an image is not on the item itself allows us to always have a value, while at the same time not taking up space in the database. As well as if we change our mind on the default image's url in the future, it's a 1 line change instead of a database update.

#INC-5312: Missing image #3
Fixes #4